requirements.txt
Flask-HTTPAuth==3.2.2
Flask-Login==0.1.3
Flask-SQLAlchemy==2.2
html5lib==0.999999999
httplib2==0.10.3
Jinja2==2.9.6
MarkupSafe==1.0
oauth2client==4.1.0
packaging==16.8
requests==2.16.5
SQLAlchemy==1.1.10
urllib3==1.21.1
webencodings==0.5.1
werkzeug==0.8.3